Added support for MCP23008 via two wire interface
This commit adds generic two wire code for reading and writing devices
on the two wire bus.

It also adds specific code for supporting an MCP23008 GPIO expander.
Two of the AVR's GPIOs are used as the reset output to reset the
MCP23008 and an interrupt input to reset it. These are defined in
the header.